% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resilience_functions.R
\docType{data}
\name{oxygen}
\alias{oxygen}
\title{Simulated Dissolved Oxygen in River Spree}
\format{A data frame with 23425 rows and 5 variables}
\usage{
data(oxygen)
}
\description{
Simulated concentrations of dissolved oxygen (DO) in mg/l for differenc
scenarios:
S2: storage as it exists in year 2020
S3: increased storage compared to the storage in year 2020
S4: reduced impervious area
S5: increased DO (what does that mean?)
}
\details{
\itemize{
  \item timestamp (POSIXct). Timestamp
  \item S2_storage_2020 (numeric). D
  \item S3_storage_increase (numeric).
  \item S4_red_Imp_Surface (numeric).
  \item S5_increase_in_DO (numeric).
}
}
\keyword{datasets}
