% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resilience_functions.R
\name{resilience.events}
\alias{resilience.events}
\title{Calculate Resilience by Failure Event}
\usage{
resilience.events(time_stamp, Pt, Pa, Pmax, evtSepTime, signalWidth)
}
\arguments{
\item{time_stamp}{vector containing timestamp (sorted in ascending order)}

\item{Pt}{vector with performance P(t) (same length as timestamp)}

\item{Pa}{accpetable performance}

\item{Pmax}{maximal failure}

\item{evtSepTime}{"event separation time" in seconds. Maximal allowed time
difference between two consecutive timestamps within the same event.}

\item{signalWidth}{"signal width" in seconds. Length of time interval that
one timestamp is representing, e.g. 5*60 = 300 if each timestamp
respresents a time interval of five minutes (as e.g. a time series is
recorded on a five minute time scale). This parameter is needed to
calculate event durations.}
}
\value{
Returns data.frame containing one row by failure event.
First columns are identical to kwb.event::hsEvents. Following columns
are additional resilience indices:
\itemize{
\item Sev: severity by event
\item Res0: resilience index by event
\item trec: recovery time in seconds
\item trec_percent: trec relative to event duration in per cent
\item worst_P: P(t) closest to Pmax within event
}
script is stopped if no failure event with message "Pa never
  exceeded"
}
\description{
Calculates resilience indices (see \href{https://www.researchgate.net/publication/326040304_Quantitative_Beschreibung_der_Resilienz_urbaner_Wassersysteme}{Matzinger et al. 2018})
for each failure event in time series of performance P(t).
Failure is defined by acceptable
performance Pa and maximal failure Pmax.
}
