% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resilience_functions.R
\name{resilience.summary}
\alias{resilience.summary}
\title{Calculate Overall Resilience Index for One or Several Performance Time Series}
\usage{
resilience.summary(time_stamp, Pt, Pa, Pmax, evtSepTime, signalWidth)
}
\arguments{
\item{time_stamp}{vector containing timestamp (sorted in ascending order)}

\item{Pt}{vector or data.frame (if several colums) with performance P(t)
(same length as timestamp)}

\item{Pa}{accpetable performance}

\item{Pmax}{maximal failure}

\item{evtSepTime}{"event separation time" in seconds. Maximal allowed time
difference between two consecutive timestamps within the same event.}

\item{signalWidth}{"signal width" in seconds. Length of time interval that
one timestamp is representing, e.g. 5*60 = 300 if each timestamp
respresents a time interval of five minutes (as e.g. a time series is
recorded on a five minute time scale). This parameter is needed to
calculate event durations.}
}
\value{
Returns data.frame containing one row by time series. Columns are:
\itemize{
   \item num_events: number of failure events in time series
   \item worst_P: P(t) closest to Pmax within time series
   \item total_dur: total duration of failure events in seconds
   \item total_trec: total recovery time of failure events in seconds
   \item mean_trec_percent: trec relative to event duration in per cent, averaged
  over all failure events in time series
   \item Sev: severity over entire time series (=0 if no exceedance of Pa)
   \item Res0: resilience index over entire time series (=1 if no exceedance
  of Pa)
}
}
\description{
Calculates resilience indices (see \href{https://www.researchgate.net/publication/326040304_Quantitative_Beschreibung_der_Resilienz_urbaner_Wassersysteme}{Matzinger et al. 2018})
for entire time series of performance P(t).
Failure is defined by acceptable
performance Pa and maximal failure Pmax. 
Entire time series is considered
}
